refactor(orchestrator): extract routing package and standardize Docker builds
Extract routing schemas, action types, and state definitions from the
orchestrator into a new packages/routing library (@catalyst/routing).
Update orchestrator and CLI imports to use the shared package.

Standardize all Dockerfiles as multi-stage builds with full workspace
resolution. Fix workspace symlink issue where Docker COPY --from
dereferences bun workspace links by re-running bun install in the
runtime stage. Add missing Dockerfiles for orders-api and product-api.
Add GATEWAY_STARTED log marker for container test compatibility.
