+++ apps/orchestrator/package.json
@@ -13,6 +13,7 @@
     "@catalyst/auth-service": "catalog:",
     "@catalyst/authorization": "catalog:",
     "@catalyst/config": "catalog:",
+    "@catalyst/routing": "catalog:",
     "@hono/capnweb": "catalog:",
     "capnweb": "catalog:",
     "hono": "catalog:",
