+++ apps/orchestrator/src/index.ts
@@ -1,5 +1,11 @@
-import { DataChannelDefinitionSchema } from './routing/datachannel.js'
 import { CatalystNodeBus } from './orchestrator.js'
 
-export { DataChannelDefinitionSchema as ServiceDefinitionSchema }
 export { CatalystNodeBus }
+
+// Re-export routing types for backward compatibility
+export {
+  DataChannelDefinitionSchema as ServiceDefinitionSchema,
+  type PeerInfo,
+  type InternalRoute,
+  type DataChannelDefinition,
+} from '@catalyst/routing'
