+++ apps/orchestrator/src/types.ts
@@ -1,7 +1,6 @@
-import { Role } from '@catalyst/authorization'
+import type { Action, RouteTable } from '@catalyst/routing'
 import { z } from 'zod'
-import type { RouteTable } from './routing/state.js'
-import type { Action } from './schema.js'
+import { Role } from '@catalyst/authorization'
 
 import { NodeConfigSchema } from '@catalyst/config'
