+++ examples/movies-api/Dockerfile
@@ -1,18 +1,66 @@
+# Build from repo root: docker build -f examples/movies-api/Dockerfile -t catalyst-example-movies-api .
+FROM oven/bun:1.3.6-alpine AS deps
+
+WORKDIR /app
+
+# Copy root workspace config (includes catalog entries)
+COPY package.json bun.lock ./
+
+# Copy ALL workspace package.json files for complete resolution
+COPY apps/auth/package.json apps/auth/package.json
+COPY apps/cli/package.json apps/cli/package.json
+COPY apps/gateway/package.json apps/gateway/package.json
+COPY apps/node/package.json apps/node/package.json
+COPY apps/orchestrator/package.json apps/orchestrator/package.json
+COPY packages/authorization/package.json packages/authorization/package.json
+COPY packages/config/package.json packages/config/package.json
+COPY packages/routing/package.json packages/routing/package.json
+COPY packages/sdk/package.json packages/sdk/package.json
+COPY packages/telemetry/package.json packages/telemetry/package.json
+COPY packages/types/package.json packages/types/package.json
+COPY examples/books-api/package.json examples/books-api/package.json
+COPY examples/movies-api/package.json examples/movies-api/package.json
+COPY examples/orders-api/package.json examples/orders-api/package.json
+COPY examples/product-api/package.json examples/product-api/package.json
+
+# Install production dependencies
+RUN bun install --omit=dev --ignore-scripts
+
+# Runtime stage
 FROM oven/bun:1.3.6-alpine
 
 WORKDIR /app
 
-# Copy root config
+# Copy external dependencies from deps stage
+COPY --from=deps /app/node_modules ./node_modules
+
+# Copy root config + lockfile for workspace resolution
 COPY package.json bun.lock ./
 
-# Copy movies-api package dependencies
+# Copy ALL workspace package.json files for workspace resolution
+COPY apps/auth/package.json apps/auth/package.json
+COPY apps/cli/package.json apps/cli/package.json
+COPY apps/gateway/package.json apps/gateway/package.json
+COPY apps/node/package.json apps/node/package.json
+COPY apps/orchestrator/package.json apps/orchestrator/package.json
+COPY packages/authorization/package.json packages/authorization/package.json
+COPY packages/config/package.json packages/config/package.json
+COPY packages/routing/package.json packages/routing/package.json
+COPY packages/sdk/package.json packages/sdk/package.json
+COPY packages/telemetry/package.json packages/telemetry/package.json
+COPY packages/types/package.json packages/types/package.json
+COPY examples/books-api/package.json examples/books-api/package.json
 COPY examples/movies-api/package.json examples/movies-api/package.json
+COPY examples/orders-api/package.json examples/orders-api/package.json
+COPY examples/product-api/package.json examples/product-api/package.json
 
-# Install production dependencies
-RUN bun install --ignore-scripts --omit=dev
-
-# Copy movies-api package source code
+# Copy app source
 COPY examples/movies-api examples/movies-api
+
+# Recreate workspace symlinks (fast — external deps already cached)
+RUN bun install --omit=dev --ignore-scripts
+
+WORKDIR /app/examples/movies-api
 
 ENV PORT=8080
 EXPOSE 8080
@@ -20,4 +68,4 @@ EXPOSE 8080
 RUN addgroup -S appgroup && adduser -S appuser -G appgroup
 USER appuser
 
-CMD ["bun", "run", "examples/movies-api/src/index.ts"]
+CMD ["bun", "run", "src/index.ts"]
